+++ .travis.yml
@@ -17,11 +17,10 @@ script:
 
 matrix:
   include:
-  
     - os: osx
       osx_image: xcode9.3
 
     - os: linux
 
 node_js:
-  - "10"
+  - '11'
