+++ binding.gyp
@@ -12,7 +12,6 @@
         "src/cpp/v8utils/v8utils.cpp",
         "src/cpp/RoaringBitmap32/RoaringBitmap32.cpp",
         "src/cpp/RoaringBitmap32/RoaringBitmap32_operations.cpp",
-        "src/cpp/RoaringBitmap32/RoaringBitmap32_serialization.cpp",
         "src/cpp/RoaringBitmap32BufferedIterator/RoaringBitmap32BufferedIterator.cpp"
       ]
     }
@@ -45,7 +44,6 @@
               "src/cpp/v8utils/v8utils.cpp",
               "src/cpp/RoaringBitmap32/RoaringBitmap32.cpp",
               "src/cpp/RoaringBitmap32/RoaringBitmap32_operations.cpp",
-              "src/cpp/RoaringBitmap32/RoaringBitmap32_serialization.cpp",
               "src/cpp/RoaringBitmap32BufferedIterator/RoaringBitmap32BufferedIterator.cpp"
             ]
           },
@@ -71,7 +69,6 @@
               "src/cpp/v8utils/v8utils.cpp",
               "src/cpp/RoaringBitmap32/RoaringBitmap32.cpp",
               "src/cpp/RoaringBitmap32/RoaringBitmap32_operations.cpp",
-              "src/cpp/RoaringBitmap32/RoaringBitmap32_serialization.cpp",
               "src/cpp/RoaringBitmap32BufferedIterator/RoaringBitmap32BufferedIterator.cpp"
             ]
           }
