+++ package.json
@@ -64,7 +64,7 @@
     "eslint-config-quick": "^0.1.13",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jest": "^21.26.2",
-    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-node": "^8.0.0",
     "fastbitset": "^0.2.8",
     "jest": "^23.6.0",
     "physical-cpu-count": "^2.0.0",
