+++ scripts/prebuild.js
@@ -5,6 +5,7 @@ args.push(process.argv[0])
 args.push('-t', '8.0.0')
 args.push('-t', '9.0.0')
 args.push('-t', '10.0.0')
+args.push('-t', '11.0.0')
 
 if (process.env.PREBUILD_GITHUB_TOKEN) {
   args.push('-u', process.env.PREBUILD_GITHUB_TOKEN)
